Added in one more scenario
Added in Signals to Destroy GameObjects
