+++ Assets/Rewards/CookAMealLost.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ff44240d50583fd48b829bf91cd529eb
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
