+++ Assets/Rewards/CookAMealWon.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5f9999bd88e38ec40b3a59750776e12a
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
