+++ Assets/Scripts/NPC/NpcInteraction.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections;
 using System.Collections.Generic;
+using Singletons;
 using Unity.VisualScripting;
 using UnityEngine;
