+++ Assets/Scripts/Options/EventDisplay.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 using Options;
+using Singletons;
 using TMPro;
 using UnityEngine;
 using UnityEngine.UI;
