+++ Assets/Scripts/Player/PlayerMovement.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 using NPC;
+using Singletons;
 using UnityEngine;
 
 namespace Player
