+++ Assets/Scripts/Player/UIStatsManager.cs
@@ -1,5 +1,6 @@
 using System.Collections;
 using System.Collections.Generic;
+using Singletons;
 using UnityEngine;
 using UnityEngine.UI;
 using TMPro;
