+++ Assets/Scripts/Singletons.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: eeca2616fdadab64ebf63853e4f33188
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
