+++ Assets/Scripts/Singletons/GameManager.cs.meta
